research_ai/requirements.txt
arxiv
sentence-transformers
chromadb
semanticscholar
networkx
umap-learn
hdbscan
bertopic
crewai
langchain
streamlit
plotly
pyvis
markdown
reportlab
python-dotenv
pytest
langchain-openai
pandas
fpdf2
langchain-google-genai
fpdf
seaborn
